requirements.txt
pip==19.3.1
selenium==3.141.0
pylint==2.4.4
autopep8==1.4.4
pytest==5.3.4
python-dotenv==0.10.5